Simplify generators a bit and unify variadic macros
