+++ rustler/otp_headers/codegen/src/generator/forwarders.rs
@@ -1,8 +1,8 @@
-use crate::parser::{ApiArg, CType};
 use std::io::Write;
 
 use super::{
-    ApiBuilder, DONE, Res, render_arg_names, render_type_args, write_ret, write_variadic_fn_type,
+    ApiArg, ApiBuilder, CType, Res, render_arg_names, render_type_args, write_ret,
+    write_variadic_fn_type,
 };
 
 /// Builds the callable forwarder functions used when the direct-symbols
@@ -32,26 +32,11 @@ impl<W: Write> ApiBuilder for ForwardersApiBuilder<'_, W> {
 
     fn variadic_func(&mut self, ret: &CType, name: &str, args: &[ApiArg]) -> Res {
         let args_sig = render_type_args(args);
-        writeln!(self.0, "#[macro_export] macro_rules! {name} {{")?;
-        writeln!(
-            self.0,
-            "    ( $( $arg:expr ),* ) => {{ $crate::sys::get_{name}()($($arg),*) }};"
-        )?;
-        writeln!(
-            self.0,
-            "    ( $( $arg:expr ),+, ) => {{ {name}!($($arg),*) }};"
-        )?;
-        writeln!(self.0, "}}\n")?;
-        writeln!(self.0, "pub use {name};\n")?;
 
         write!(self.0, "pub unsafe fn get_{name}() -> ")?;
         write_variadic_fn_type(self.0, &args_sig, ret)?;
         writeln!(self.0, " {{")?;
         writeln!(self.0, "    DYN_NIF_CALLBACKS.{name}.unwrap_unchecked()")?;
         writeln!(self.0, "}}\n")
-    }
-
-    fn dummy(&mut self, _name: &str) -> Res {
-        DONE
     }
 }
