+++ rustler/otp_headers/nif_version_2_15/api.4.direct.rs
@@ -147,16 +147,28 @@ extern "C" {
     pub fn enif_term_type(env: *mut ErlNifEnv, term: ERL_NIF_TERM) -> ErlNifTermType;
 }
 
+pub unsafe fn get_enif_make_tuple() -> unsafe extern "C" fn (env: *mut ErlNifEnv, cnt: c_uint, ...) -> ERL_NIF_TERM {
+    std::mem::transmute(__variadic_enif_make_tuple as *const ())
+}
+
+pub unsafe fn get_enif_make_list() -> unsafe extern "C" fn (env: *mut ErlNifEnv, cnt: c_uint, ...) -> ERL_NIF_TERM {
+    std::mem::transmute(__variadic_enif_make_list as *const ())
+}
+
+pub unsafe fn get_enif_fprintf() -> unsafe extern "C" fn (filep: *mut c_void, format: *const c_char, ...) -> c_int {
+    std::mem::transmute(__variadic_enif_fprintf as *const ())
+}
+
+pub unsafe fn get_enif_snprintf() -> unsafe extern "C" fn (buffer: *mut c_char, size: size_t, format: *const c_char, ...) -> c_int {
+    std::mem::transmute(__variadic_enif_snprintf as *const ())
+}
+
 #[macro_export] macro_rules! enif_make_tuple {
     ( $( $arg:expr ),* ) => { $crate::sys::get_enif_make_tuple()($($arg),*) };
     ( $( $arg:expr ),+, ) => { enif_make_tuple!($($arg),*) };
 }
 
 pub use enif_make_tuple;
-
-pub unsafe fn get_enif_make_tuple() -> unsafe extern "C" fn (env: *mut ErlNifEnv, cnt: c_uint, ...) -> ERL_NIF_TERM {
-    std::mem::transmute(__variadic_enif_make_tuple as *const ())
-}
 
 #[macro_export] macro_rules! enif_make_list {
     ( $( $arg:expr ),* ) => { $crate::sys::get_enif_make_list()($($arg),*) };
@@ -165,10 +177,6 @@ pub unsafe fn get_enif_make_tuple() -> unsafe extern "C" fn (env: *mut ErlNifEnv
 
 pub use enif_make_list;
 
-pub unsafe fn get_enif_make_list() -> unsafe extern "C" fn (env: *mut ErlNifEnv, cnt: c_uint, ...) -> ERL_NIF_TERM {
-    std::mem::transmute(__variadic_enif_make_list as *const ())
-}
-
 #[macro_export] macro_rules! enif_fprintf {
     ( $( $arg:expr ),* ) => { $crate::sys::get_enif_fprintf()($($arg),*) };
     ( $( $arg:expr ),+, ) => { enif_fprintf!($($arg),*) };
@@ -176,18 +184,10 @@ pub unsafe fn get_enif_make_list() -> unsafe extern "C" fn (env: *mut ErlNifEnv,
 
 pub use enif_fprintf;
 
-pub unsafe fn get_enif_fprintf() -> unsafe extern "C" fn (filep: *mut c_void, format: *const c_char, ...) -> c_int {
-    std::mem::transmute(__variadic_enif_fprintf as *const ())
-}
-
 #[macro_export] macro_rules! enif_snprintf {
     ( $( $arg:expr ),* ) => { $crate::sys::get_enif_snprintf()($($arg),*) };
     ( $( $arg:expr ),+, ) => { enif_snprintf!($($arg),*) };
 }
 
 pub use enif_snprintf;
-
-pub unsafe fn get_enif_snprintf() -> unsafe extern "C" fn (buffer: *mut c_char, size: size_t, format: *const c_char, ...) -> c_int {
-    std::mem::transmute(__variadic_enif_snprintf as *const ())
-}
