+++ rustler/otp_headers/nif_version_2_18/api.8.direct.rs
@@ -153,16 +153,32 @@ extern "C" {
     pub fn enif_max_atom_cache_index() -> c_uint;
 }
 
+pub unsafe fn get_enif_make_tuple() -> unsafe extern "C" fn (env: *mut ErlNifEnv, cnt: c_uint, ...) -> ERL_NIF_TERM {
+    std::mem::transmute(__variadic_enif_make_tuple as *const ())
+}
+
+pub unsafe fn get_enif_make_list() -> unsafe extern "C" fn (env: *mut ErlNifEnv, cnt: c_uint, ...) -> ERL_NIF_TERM {
+    std::mem::transmute(__variadic_enif_make_list as *const ())
+}
+
+pub unsafe fn get_enif_fprintf() -> unsafe extern "C" fn (filep: *mut c_void, format: *const c_char, ...) -> c_int {
+    std::mem::transmute(__variadic_enif_fprintf as *const ())
+}
+
+pub unsafe fn get_enif_snprintf() -> unsafe extern "C" fn (buffer: *mut c_char, size: size_t, format: *const c_char, ...) -> c_int {
+    std::mem::transmute(__variadic_enif_snprintf as *const ())
+}
+
+pub unsafe fn get_enif_set_option() -> unsafe extern "C" fn (env: *mut ErlNifEnv, opt: ErlNifOption, ...) -> c_int {
+    std::mem::transmute(__variadic_enif_set_option as *const ())
+}
+
 #[macro_export] macro_rules! enif_make_tuple {
     ( $( $arg:expr ),* ) => { $crate::sys::get_enif_make_tuple()($($arg),*) };
     ( $( $arg:expr ),+, ) => { enif_make_tuple!($($arg),*) };
 }
 
 pub use enif_make_tuple;
-
-pub unsafe fn get_enif_make_tuple() -> unsafe extern "C" fn (env: *mut ErlNifEnv, cnt: c_uint, ...) -> ERL_NIF_TERM {
-    std::mem::transmute(__variadic_enif_make_tuple as *const ())
-}
 
 #[macro_export] macro_rules! enif_make_list {
     ( $( $arg:expr ),* ) => { $crate::sys::get_enif_make_list()($($arg),*) };
@@ -171,20 +187,12 @@ pub unsafe fn get_enif_make_tuple() -> unsafe extern "C" fn (env: *mut ErlNifEnv
 
 pub use enif_make_list;
 
-pub unsafe fn get_enif_make_list() -> unsafe extern "C" fn (env: *mut ErlNifEnv, cnt: c_uint, ...) -> ERL_NIF_TERM {
-    std::mem::transmute(__variadic_enif_make_list as *const ())
-}
-
 #[macro_export] macro_rules! enif_fprintf {
     ( $( $arg:expr ),* ) => { $crate::sys::get_enif_fprintf()($($arg),*) };
     ( $( $arg:expr ),+, ) => { enif_fprintf!($($arg),*) };
 }
 
 pub use enif_fprintf;
-
-pub unsafe fn get_enif_fprintf() -> unsafe extern "C" fn (filep: *mut c_void, format: *const c_char, ...) -> c_int {
-    std::mem::transmute(__variadic_enif_fprintf as *const ())
-}
 
 #[macro_export] macro_rules! enif_snprintf {
     ( $( $arg:expr ),* ) => { $crate::sys::get_enif_snprintf()($($arg),*) };
@@ -193,20 +201,12 @@ pub unsafe fn get_enif_fprintf() -> unsafe extern "C" fn (filep: *mut c_void, fo
 
 pub use enif_snprintf;
 
-pub unsafe fn get_enif_snprintf() -> unsafe extern "C" fn (buffer: *mut c_char, size: size_t, format: *const c_char, ...) -> c_int {
-    std::mem::transmute(__variadic_enif_snprintf as *const ())
-}
-
 #[macro_export] macro_rules! enif_set_option {
     ( $( $arg:expr ),* ) => { $crate::sys::get_enif_set_option()($($arg),*) };
     ( $( $arg:expr ),+, ) => { enif_set_option!($($arg),*) };
 }
 
 pub use enif_set_option;
-
-pub unsafe fn get_enif_set_option() -> unsafe extern "C" fn (env: *mut ErlNifEnv, opt: ErlNifOption, ...) -> c_int {
-    std::mem::transmute(__variadic_enif_set_option as *const ())
-}
 
 
 /// See [enif_make_int64](http://www.erlang.org/doc/man/erl_nif.html#enif_make_int64) at erlang.org
